Refactoring getUserManagementAccess function
