+++ test/services/getUserManagementAccess.test.ts
@@ -1,14 +1,12 @@
-import getUserManagementAccess, { AuthenticationTokenPayload } from "./getUserManagementAccess"
+import getUserManagementAccess, { AuthenticationTokenPayload } from "../../src/services/getUserManagementAccess"
 
 it("should say that user has only access to User Management when user has one of the User Management's user groups", () => {
-  const result = getUserManagementAccess({ groups: ["UserManager"] } as AuthenticationTokenPayload)
-
-  const { hasAccessToUserManagement } = result
-  expect(hasAccessToUserManagement).toBe(true)
+  const hasAccess = getUserManagementAccess({ groups: ["UserManager"] } as AuthenticationTokenPayload)
+  expect(hasAccess).toBe(true)
 })
 
-it("should say it does not have access to User Management when user does not have one of the User Management's user groups", () => {
-  const result = getUserManagementAccess({
+it("should check if it doesn't have access to the UserManagement page", () => {
+  const hasAccess = getUserManagementAccess({
     groups: [
       "GeneralHandler",
       "Allocator",
@@ -23,6 +21,12 @@ it("should say it does not have access to User Management when user does not hav
     ]
   } as AuthenticationTokenPayload)
 
-  const { hasAccessToUserManagement } = result
-  expect(hasAccessToUserManagement).toBe(false)
+  expect(hasAccess).toBe(false)
+})
+
+it("should say it does not have access to User Management when there are no user groups", () => {
+  const hasAccess = getUserManagementAccess({
+    groups: ["IncorrectGroupName"]
+  } as unknown as AuthenticationTokenPayload)
+  expect(hasAccess).toBe(false)
 })
